Code cleanup: Split classes into new files Normalised code style Added XML comments to publicly exposed items (still WIP) Abstracted json property names to have more c#-ish names in the objects

+++ Tweety/Webhooks/SubscriptionsManager.cs
@@ -1,77 +1,34 @@
 ﻿using Newtonsoft.Json;
 using System;
-using System.Collections.Generic;
 using System.Net;
 using System.Net.Http;
-using System.Text;
 using System.Threading.Tasks;
 using Tweety.Authentication;
-using Tweety.Models;
 using Tweety.Models.Twitter;
 
 namespace Tweety.Webhooks
 {
-
     /// <summary>
     /// This class will help in managing the webhook subscription.
     /// </summary>
     public class SubscriptionsManager
     {
-        public TweetyAuthContext AuthContext { get; set; }
-
+        /// <summary>
+        /// Initializes a new instance of the <see cref="SubscriptionsManager"/> class.
+        /// </summary>
+        /// <param name="context">The context.</param>
         public SubscriptionsManager(TweetyAuthContext context)
         {
             AuthContext = context;
         }
 
-
         /// <summary>
-        /// Subscribe current user (from the auth context) to a webhook by Id.
+        /// Gets or sets the authentication context.
         /// </summary>
-        /// <param name="webhookId">Webhook Id to subscribe to.</param>
-        /// <returns>true indicates successfull subscribtion.</returns>
-        public async Task<Result<bool>> Subscribe(string webhookId)
-        {
-
-            if (string.IsNullOrEmpty(webhookId))
-            {
-                throw new ArgumentException(nameof(webhookId));
-            }
-
-            //TODO: Provide a generic class to make Twitter API Requests.
-            string resourceUrl = $"https://api.twitter.com/1.1/account_activity/webhooks/{webhookId}/subscriptions.json";
-
-            HttpResponseMessage response;
-            using (HttpClient client = new HttpClient())
-            {
-
-                client.DefaultRequestHeaders.Add("Authorization", AuthHeaderBuilder.Build(AuthContext, HttpMethod.Post, resourceUrl));
-
-                response = await client.PostAsync(resourceUrl, new StringContent(""));
-            }
-
-
-            if (response.StatusCode == HttpStatusCode.NoContent)
-            {
-                return new Result<bool>(true);
-            }
-
-            string jsonResponse = await response.Content.ReadAsStringAsync();
-
-            if (!string.IsNullOrEmpty(jsonResponse))
-            {
-                TwitterError err = JsonConvert.DeserializeObject<TwitterError>(jsonResponse);
-                return new Result<bool>(err);
-            }
-            else
-            {
-                //TODO: Provide a way to return httpstatus code 
-
-                return new Result<bool>();
-            }
-
-        }
-
+        /// <value>
+        /// The authentication context.
+        /// </value>
+        public TweetyAuthContext AuthContext { get; set; }
         /// <summary>
         /// Checks if the current user (from the auth context) is subscribed to a webhook by Id.
         /// </summary>
@@ -79,7 +36,6 @@ namespace Tweety.Webhooks
         /// <returns>true indicates existed subscribtion.</returns>
         public async Task<Result<bool>> CheckSubscription(string webhookId)
         {
-
             if (string.IsNullOrEmpty(webhookId))
             {
                 throw new ArgumentException(nameof(webhookId));
@@ -91,7 +47,6 @@ namespace Tweety.Webhooks
             HttpResponseMessage response;
             using (HttpClient client = new HttpClient())
             {
-
                 client.DefaultRequestHeaders.Add("Authorization", AuthHeaderBuilder.Build(AuthContext, HttpMethod.Get, resourceUrl));
 
                 response = await client.GetAsync(resourceUrl);
@@ -107,7 +62,7 @@ namespace Tweety.Webhooks
             if (!string.IsNullOrEmpty(jsonResponse))
             {
                 TwitterError err = JsonConvert.DeserializeObject<TwitterError>(jsonResponse);
-                if(err.Errors.Count == 1 && err.Errors[0].Code == 34)
+                if (err.Errors.Count == 1 && err.Errors[0].Code == 34)
                 {
                     // Twitter API will return : {"code":34,"message":"Sorry, that page does not exist."} if you try to check a webhook with 0 subscribers,
                     // Which means, you're not subscribed.
@@ -118,23 +73,18 @@ namespace Tweety.Webhooks
             }
             else
             {
-
-                //TODO: Provide a way to return httpstatus code 
+                //TODO: Provide a way to return httpstatus code
                 return new Result<bool>();
             }
-
-
         }
 
-
         /// <summary>
-        /// Unsubscribe current user (from the auth context) from a webhook by Id.
+        /// Subscribe current user (from the auth context) to a webhook by Id.
         /// </summary>
-        /// <param name="webhookId">Webhook Id to unsubscribe from.</param>
-        /// <returns>true indicates successful deletion.</returns>
-        public async Task<Result<bool>> Unsubscribe(string webhookId)
+        /// <param name="webhookId">Webhook Id to subscribe to.</param>
+        /// <returns>true indicates successfull subscribtion.</returns>
+        public async Task<Result<bool>> Subscribe(string webhookId)
         {
-
             if (string.IsNullOrEmpty(webhookId))
             {
                 throw new ArgumentException(nameof(webhookId));
@@ -146,7 +96,48 @@ namespace Tweety.Webhooks
             HttpResponseMessage response;
             using (HttpClient client = new HttpClient())
             {
+                client.DefaultRequestHeaders.Add("Authorization", AuthHeaderBuilder.Build(AuthContext, HttpMethod.Post, resourceUrl));
 
+                response = await client.PostAsync(resourceUrl, new StringContent(""));
+            }
+
+            if (response.StatusCode == HttpStatusCode.NoContent)
+            {
+                return new Result<bool>(true);
+            }
+
+            string jsonResponse = await response.Content.ReadAsStringAsync();
+
+            if (!string.IsNullOrEmpty(jsonResponse))
+            {
+                TwitterError err = JsonConvert.DeserializeObject<TwitterError>(jsonResponse);
+                return new Result<bool>(err);
+            }
+            else
+            {
+                //TODO: Provide a way to return httpstatus code
+
+                return new Result<bool>();
+            }
+        }
+        /// <summary>
+        /// Unsubscribe current user (from the auth context) from a webhook by Id.
+        /// </summary>
+        /// <param name="webhookId">Webhook Id to unsubscribe from.</param>
+        /// <returns>true indicates successful deletion.</returns>
+        public async Task<Result<bool>> Unsubscribe(string webhookId)
+        {
+            if (string.IsNullOrEmpty(webhookId))
+            {
+                throw new ArgumentException(nameof(webhookId));
+            }
+
+            //TODO: Provide a generic class to make Twitter API Requests.
+            string resourceUrl = $"https://api.twitter.com/1.1/account_activity/webhooks/{webhookId}/subscriptions.json";
+
+            HttpResponseMessage response;
+            using (HttpClient client = new HttpClient())
+            {
                 client.DefaultRequestHeaders.Add("Authorization", AuthHeaderBuilder.Build(AuthContext, HttpMethod.Delete, resourceUrl));
 
                 response = await client.DeleteAsync(resourceUrl);
@@ -173,14 +164,9 @@ namespace Tweety.Webhooks
             }
             else
             {
-
-                //TODO: Provide a way to return httpstatus code 
+                //TODO: Provide a way to return httpstatus code
                 return new Result<bool>();
             }
-
-
         }
-
-
     }
 }
